requirements.txt
asgiref==3.7.2
Django==5.0.1
Faker==22.6.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
psycopg2==2.9.9
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pillow==10.2.0
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
flake8==7.0.0
pytest-cov==5.0.0
